fix load data issues due to transition animation

+++ apps/dotcms-ui/src/app/portlets/dot-edit-page/components/dot-palette/dot-palette-contentlets/dot-palette-contentlets.component.ts
@@ -1,43 +1,46 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import {
+    Component,
+    EventEmitter,
+    Input,
+    OnChanges,
+    Output,
+    SimpleChanges,
+    ViewChild
+} from '@angular/core';
 import { DotContentletEditorService } from '@dotcms/app/view/components/dot-contentlet-editor/services/dot-contentlet-editor.service';
-// import { CoreWebService } from '@dotcms/dotcms-js';
 import { DotCMSContentlet } from '@dotcms/dotcms-models';
 import { DotPaginatorESContentService } from '@services/dot-paginator-es-content/dot-paginator-es-content.service';
-// import { th } from 'date-fns/locale';
 import { LazyLoadEvent } from 'primeng/api';
 import { take } from 'rxjs/operators';
-// import { DotCMSContentType } from '@dotcms/dotcms-models';
-// import { Subject } from 'rxjs';
-// import { DotContentletEditorService } from '@components/dot-contentlet-editor/services/dot-contentlet-editor.service';
+import { DotPaletteInputFilterComponent } from '../dot-palette-input-filter/dot-palette-input-filter.component';
 
 @Component({
     selector: 'dot-palette-contentlets',
     templateUrl: './dot-palette-contentlets.component.html',
     styleUrls: ['./dot-palette-contentlets.component.scss']
 })
-export class DotPaletteContentletsComponent implements OnInit {
-    // @Input() items: DotCMSContentType[] = [];
-    filter: string;
+export class DotPaletteContentletsComponent implements OnChanges {
+    @ViewChild('filterInput', { static: true })
+    filterInput: DotPaletteInputFilterComponent;
+
     @Input() contentTypeVariable: string;
     @Output() hide = new EventEmitter();
     items: DotCMSContentlet[];
     hideNoResults = true;
 
-    // private destroy$: Subject<boolean> = new Subject<boolean>();
-
     constructor(
         public paginatorESService: DotPaginatorESContentService,
-        // private coreWebService: CoreWebService,
         private dotContentletEditorService: DotContentletEditorService
     ) {}
 
-    ngOnInit() {
-        // this.paginatorESService.url = `/api/content/render/false/query/+contentType:${this.contentTypeVariable}`;
-        this.paginatorESService.setExtraParams('+contentType', this.contentTypeVariable);
-        this.paginatorESService.paginationPerPage = 15;
-        this.paginatorESService.deleteExtraParams('+title');
+    ngOnChanges(changes: SimpleChanges): void {
+        if (changes?.contentTypeVariable?.currentValue) {
+            this.paginatorESService.setExtraParams('+contentType', this.contentTypeVariable);
+            this.paginatorESService.deleteExtraParams('+title');
+            this.paginatorESService.paginationPerPage = 15;
 
-        this.loadData();
+            this.loadData();
+        }
     }
 
     /**
@@ -47,46 +50,23 @@ export class DotPaletteContentletsComponent implements OnInit {
      * @memberof DotAppsConfigurationComponent
      */
     loadData(event?: LazyLoadEvent): void {
-        console.log(this.hideNoResults)
         this.paginatorESService
             .getWithOffset((event && event.first) || 0)
             .pipe(take(1))
             .subscribe((data: any[]) => {
-                console.log('****data', data, this.paginatorESService.totalRecords);
                 this.items = data;
                 this.hideNoResults = !!data?.length;
-                console.log(this.hideNoResults)
-
-                // const app = [].concat(apps)[0];
-                // this.apps.sites = event ? this.apps.sites.concat(app.sites) : app.sites;
-                // this.apps.configurationsCount = app.configurationsCount;
-                // this.totalRecords = this.paginationService.totalRecords;
-                // this.hideLoadDataButton = !this.isThereMoreData(this.apps.sites.length);
             });
 
-        // TODO: remove later
-        // this.coreWebService
-        //     .requestView({
-        //         url: this.paginatorService.url
-        //     })
-        //     .pipe(pluck('contentlets'))
-        //     .subscribe((data) => {
-        //         console.log('==== data', data);
-        //         this.items = data
-
-        //     });
     }
 
     paginate(event: LazyLoadEvent) {
-        console.log('**event', event);
         this.loadData(event);
-        //event.first = Index of the first record
-        //event.rows = Number of rows to display in new page
-        //event.page = Index of the new page
-        //event.pageCount = Total number of pages
     }
 
     showContentTypesList(): void {
+        this.items = null;
+        this.filterInput.searchInput.nativeElement.value = '';
         this.hide.emit();
     }
 
@@ -101,17 +81,12 @@ export class DotPaletteContentletsComponent implements OnInit {
     }
 
     filterContentlets(value: string): void {
-        console.log('**filter', value);
         value = value.trim();
         if (value.indexOf(' ') > 0) {
-            value = `'${value.replace(/'/g, '\\\'')}'`;
+            value = `'${value.replace(/'/g, "\\'")}'`;
         }
         this.paginatorESService.setExtraParams('+title', `${value}*`);
         this.loadData();
     }
 
-    // ngOnDestroy(): void {
-    //     this.destroy$.next(true);
-    //     this.destroy$.complete();
-    // }
 }
